+++ src/utilities/prTitleParsers.spec.ts
@@ -30,6 +30,21 @@ describe('checkPullRequestTitleForMergePreset', (): void => {
     );
   });
 
+  describe('given containing major bump and directory path', (): void => {
+    const title = 'bump @types/jest from 26.0.12 to 27.0.13 in /directory';
+
+    it.each(Object.values(inputParsers.AllowedMergePresets))(
+      'returns false',
+      (mergeCategory: inputParsers.AllowedMergePresets): void => {
+        expect.assertions(1);
+
+        parseInputMergePresetSpy.mockReturnValueOnce(mergeCategory);
+
+        expect(checkPullRequestTitleForMergePreset(title)).toStrictEqual(false);
+      },
+    );
+  });
+
   describe('given title containing minor bump', (): void => {
     const title = 'bump @types/jest from 26.0.12 to 26.1.0';
 
@@ -54,8 +69,47 @@ describe('checkPullRequestTitleForMergePreset', (): void => {
     });
   });
 
+  describe('given title containing minor bump and directory path', (): void => {
+    const title = 'bump @types/jest from 26.0.12 to 26.1.0 in /directory';
+
+    it('returns true for DEPENDABOT_MINOR', (): void => {
+      expect.assertions(1);
+
+      parseInputMergePresetSpy.mockReturnValueOnce(
+        inputParsers.AllowedMergePresets.DEPENDABOT_MINOR,
+      );
+
+      expect(checkPullRequestTitleForMergePreset(title)).toStrictEqual(true);
+    });
+
+    it('returns false for DEPENDABOT_PATCH', (): void => {
+      expect.assertions(1);
+
+      parseInputMergePresetSpy.mockReturnValueOnce(
+        inputParsers.AllowedMergePresets.DEPENDABOT_PATCH,
+      );
+
+      expect(checkPullRequestTitleForMergePreset(title)).toStrictEqual(false);
+    });
+  });
+
   describe('given title containing patch bump', (): void => {
     const title = 'bump @types/jest from 26.0.12 to 26.0.13';
+
+    it.each(Object.values(inputParsers.AllowedMergePresets))(
+      'returns true',
+      (mergeCategory: inputParsers.AllowedMergePresets): void => {
+        expect.assertions(1);
+
+        parseInputMergePresetSpy.mockReturnValueOnce(mergeCategory);
+
+        expect(checkPullRequestTitleForMergePreset(title)).toStrictEqual(true);
+      },
+    );
+  });
+
+  describe('given title containing patch bump and directory path', (): void => {
+    const title = 'bump @types/jest from 26.0.12 to 26.0.13 in /directory';
 
     it.each(Object.values(inputParsers.AllowedMergePresets))(
       'returns true',